return individual answers-core request from executes()
For the Autocomplete component, it's important that one
SearchBar's Autocomplete does not cause another SearchBar's
autocomplete to start displaying results. Unfortunately, this
is what happens when component's listen to a shared autocomplete
state.

Returning the request itself, in addition to dispatching an action,
lets components decide whether they would like to engage in a "push"
or "pull" relationship with answers-headless.

TEST=manual

used this in the react sample app