add select/unselectFacetOption
This commit replaces setFacets and setDisplayableFacets with
select/unselectFacetOption. This should be a more sane interface
for facets, where the main use case is a user checking/unchecking
a checkbox.

It also removes the triple duplication of the facets state, and
puts it into a single DisplayableFacets state.

TEST=auto

integration tests for selecting/unselecting facets and edge cases