add location + search intents support
This commit adds support for setting location on searches, and
receiving locationBias data from search responses.

It also adds searchIntents state to state.query. I put it on query,
because the search intent is the "search intent for the query" and can
be set by either running a search or triggering autocomplete.

J=SLAP-1576
TEST=auto