Add UniversalLimit support (#34)
Add UniversalLimit support

Delete `setLimit` and create a `setVerticalLimit` and a `setUniversalLimit`.

`setVertialLimit` takes a string and applies to the current vertical. `setUniversalLimit` takes in a `UniversalLimit` object and can affect multiple verticals on universal search.

J=SLAP-1561
TEST=manual, auto

In the sample app, try calling `setVerticalLimit` and `setUniversalLimit` and see the searches properly limit the number of results. Add unit and integration tests.