Rename StatefulCore to AnswersHeadless (#37)
Rename StatefulCore to AnswersHeadless

StatefulCore was the original prototype name which should be updated to the new name of AnswersHeadless. I also update `statefulCore` to be simply `answers` rather than `answersHeadless`. It's more concise, and I think it makes sense at this high level of abstraction. It leads to code such as `answers.executeVerticalQuery()`

This PR also creates a HeadlessConfig so that the library isn't tied to the exact same config as the one in AnswersCore

J=SLAP-1649
TEST=none

Update tests and confirm they still pass