export * as @yext/answers-core
redux-toolkit also does this with redux.
Prior to this, the recommendation was to rely on the @yext/answers-core
transitive dependency. It is not advisiable to rely on transitive dependencies,
which may even be handled differently by different bundlers.
Relying on the transitive dep  also caused issues during local development of answers-headless and answers-headless-react.

J=SLAP-1664
TEST=manual

hooked up my local answers-headless to answers-headless-react
tested out what would happen if you try to export two symbols with the same name - typescript will yell at you