Support for multiple, isolated answers headless instances (#38)
This commit adds support for multiple answers headless instances, with a new `headlessId` in `provideAnswersHeadless`.

It does so by refactoring all slice ducks into "create" methods that take in an optional namespace.
This namespace is used to scope reducers scoped to a single `State` subtree. 
In order to do this, reducers are now managed by a HeadlessReducerManager class, which handles adding additional reducers for newly added instances, and recombining the reducer map.

In order for multiple StateManagers to share the same store, the store's instantiation was pulled out of ReduxStateManager. 
_(note: This has the nice side effect of fixing a buggy interaction we currently have with redux devtools in the answers-headless-react sample app, where two instances of AnswersHeadless would appear in the devtools drop down.)_

All headless instances have their SessionTrackingState linked together, as a requirement from product.
In the future, we can work on an interface to let people easily link together instances, or even just a flag
to turn off the default linking.

I also changed the dev tools display name to AnswersHeadless, instead of the default, which is whatever the current `window.title` is.

J=SLAP-1647
TEST=manual,auto

tested multiple instances of AnswersHeadless on the same page, and setting different parts of the query in multiple child AnswersHeadless instances
test the set-state action
test linking multiple AnswersHeadless instances together using addListener()