Top level direct answer slice (#43)
Create a slice for direct answers at the top level of the state

I originally tried using the direct answer models as the state of the slice, however I was running into problems with redux when attempting it. I was able to get it working by putting the direct answer under a prop which I called 'result'. I believe that putting the direct answer under a prop is the best way of maintaining the typescript discriminating union. I think It reads well, and it gives us flexibility in case we need to add anything to the direct answer state which is specific to headless.

J=SLAP-1664
TEST=auto

Add unit test which confirms the direct answer is set properly in this new slice