Fix bug that prevented a single Filter from being unselected.
I noticed that when I had a single Filter selected and I attempted
to unselect it, it was still applied to the search. I tracked down the
cause of this issue to the `transformFiltersToCoreFormat` method. This
method incorrectly assumes that if the list of `SelectableFilter`s is length
1, the filter must be selected. That is not the case, we still need to check
the `selected` prop.

TEST=auto

I added a unit test to prevent any regressions.