Add API Extractor (#63)
Add API extractor to get an API Report, a d.ts bundle, and markdown docs, as in Answers Core and Jambo. Also, add more exports, release tags, and link to markdown docs in README. Rename `Range` to `BoundedRange` to avoid conflict with the Typescript `Range` interface.

Notes:
- When reviewing, you can ignore the docs files
- As in Answers Core, there are separate `build-ci` and `api-extractor-ci` scripts, which exit with non-zero exit codes for warning and outdated api-reports
- As in Answers Core, the automatic tests will error out if the docs are out of date
- @yext/answers-core is added as a bundled package in the config for API extractor to correctly resolve links. This adds types/interfaces from answers-core in the d.ts bundle so that the referenced docs are not treated as coming from an external package

J=SLAP-1738
TEST=manual

Ran the api extractor and inspected the api report, type bundle, and markdown docs. Made changes to the docs and checked that the docs validation script worked as expected.