Expose the various Services in Core.
This PR exposes the various Services `interfaces` in Core. It also
makes the `AnswersHeadless` constructor public. This allows for someone
to create an instance of Headless outside our `provideHeadless` facade.
It may be the case that someone wants to use their own state manager or
a custom Core instance.

J=SLAP-1928
TEST=none