Answers Agent Support (#86)
With each request, pass the additional HTTP headers. Currently, this only allows the Client-SDK header, which includes the ANSWERS_HEADLESS agent specifying the Headless version used by the site, as well as any additional agents passed to `provideAnswersHeadless`. Add unit tests to confirm this behavior.

The `provideAnswersHeadless` function is overloaded so we can pass the headers when using the internal signature but not expose the param publicly.

Like in Core, because we are importing package.json to get the version, in order to get the `commonjs` and `esm` imports to work correctly, the `outDir` was changed from `lib` to `dist`.

J=SLAP-1950
TEST=auto

Check that Jest tests pass.