Add Filter and Facet NumberRangeValue support (#88)
Add support for facets and filters with a NumberRangeValue 

Add the new models to the re-exports file.

The one issue I noticed during testing is that our filter equality logic was doing a regular `===` check to see if two filters/facet values were the same. This caused some incorrect behavior because the `===` check would return false if the two NumberRangeValues weren't the exact same object, and they usually aren't. I updated to use lodash's `isEqual` to fix the issue by performing a deep equality check. Using lodash means that the equality check should work even if the filter value is modified in the future.

J=SLAP-1980
TEST=manual, auto

Manually test setting a filter with a NumberRangeValue with `answers.setFilterOption` in the node test site and see the expected state changes. Also manually test setting a facet with `answers.setFacetOption` in the node test site as well and see the expected state changes. Add unit tests.