export SandboxEndpoints from answers-core + other new exports
This PR was made for the purpose of exporting the new SandboxEndpoints
constant from answers-core.

This new version of answers-core has a few other references types that
need to be exported, otherwise they show up as undocumented in the api-documentor
docs.

J=SLAP-2106
TEST=manual

check that documentation has no missing exports