Expand static filters (#130)
This PR makes the static filters state more flexible, so users can store more complex static filters as well as field value filters. It incorporates the changes in the Core models, including the new `StaticFilter` interface. The `SelectableFilter` interface is replaced with `SelectableStaticFilter` and the static filters state now accepts an array of the latter type. The Jest tests were updated to reflect these changes and to test that the new functionality works as expected.

J=SLAP-2326
TEST=auto, manual

See that the new and existing Jest tests pass. From the Node test-site, store static filters in State and run vertical searches. See that they execute successfully and give the correct results.